Add padding parameter
The following update successfully employs a generic version of obtaining precinct names within a given county file, based on a specified padding parameter.

This fails however with regard to collecting the yes/no vote breakdown, currently. I am working on trying to figure out the relative position of yes votes and no votes, however the current method shown in this does not work.