Minimize parameters to collect yes/no votes.
There are now two approaches to collecting yes/no votes.
The second approach was what appeared in previous commits. That
depended on additional parameters like the yes_vote_padding
to determine a yes vote's relative position in a file.

The first approach requires, in addition to the file
padding between precincts, the number of column numbers. This
approach keeps track of the colum numbers. You can read about
the strategy in ScriptsAndData/breakdown_scripts/generic.py.
This approach seems to work, with the outlet I've collected. I have
not had the chance to verify it though.