Collect yes/no votes with padding and column parameter
This commit finishes Approach #1 in breakdown_scripts/generic.py. The script correctly processes Ashland.txt, given a padding and column parameter. Note that the last item in the yes/no votes print statements are the TOTAL votes. They will need to be removed in a future commit.

For next commits, I think it would be best to try this approach with other files that follow the same padding/parameter dependency. It could make it relatively easy to gather all the data in these type of files with minimal and clear inputs.