Create generic version for processing county election PDFs
There are three changes within this commit.

(1) archived_vote_processing/
This directory holds previous methods of processing county election data on a non-generic basis.

(2) generic_breakdown.py
This contains a generic method to breakdown county PDFs with a "county" object and additional functions.

NOTE: I have not tested this script yet.

(3) instanced_version_generic.py
This provides an idea of testing the ideal functionality of
"generic_breakdown.py". This script works for the given county in that script.

What needs to be done to finish this:
Both (2) and (3) require a method of sorting through the processed county election data. Each provide the given yes/no votes for each section within a county PDF. The user simply needs to determine which votes occur during which sections and then store those into a CSV file.